Reusable component for Usersnap focus logic
