+++ src/apps/project-settings/GetPlugins/PluginGallery.tsx
@@ -3,6 +3,7 @@ import type { Plugin, PluginInstallationConfig } from '@recogito/studio-sdk';
 import { VisuallyHidden } from '@radix-ui/react-visually-hidden';
 import { PluginGalleryItem } from './PluginGalleryItem';
 import type { Project, Translations } from 'src/Types';
+import { DialogContent } from '@components/DialogContent';
 
 import './PluginGallery.css';
 
@@ -37,7 +38,7 @@ export const PluginGallery = (props: PluginGalleryProps) => {
       <Dialog.Portal>
         <Dialog.Overlay className="dialog-overlay" />
 
-        <Dialog.Content 
+        <DialogContent 
           className="dialog-content plugin-gallery">
 
           <Dialog.Title asChild>
@@ -63,7 +64,7 @@ export const PluginGallery = (props: PluginGalleryProps) => {
               </li>
             ))}
           </ul>
-        </Dialog.Content>
+        </DialogContent>
       </Dialog.Portal>
     </Dialog.Root>
   )
