+++ src/components/MetadataModal/MetadataModal.tsx
@@ -8,6 +8,7 @@ import { updateDocumentMetadata } from '@backend/crud';
 import { Button } from '@components/Button';
 import type { Document, Translations } from 'src/Types';
 import type { LibraryDocument } from '@components/DocumentLibrary';
+import { DialogContent } from '@components/DialogContent';
 import './MetadataModal.css';
 
 interface Item {
@@ -112,7 +113,7 @@ export const MetadataModal = (props: MetadataModalProps) => {
       <Dialog.Portal>
         <Dialog.Overlay className='dialog-overlay' />
 
-        <Dialog.Content className='dialog-content metadata-modal'>
+        <DialogContent className='dialog-content metadata-modal'>
           <Dialog.Title className='dialog-title'>
             <DocumentTitle
               onChange={(value) => setTitle(value)}
@@ -224,7 +225,7 @@ export const MetadataModal = (props: MetadataModalProps) => {
               <X size={16} />
             </button>
           </Dialog.Close>
-        </Dialog.Content>
+        </DialogContent>
       </Dialog.Portal>
     </Dialog.Root>
   );
